Notify the mobile app native code that the user is logged in
This is useful, because the WebView has an httpOnly=true cookie (not
readable from the native code) that is automatically set by the backend
when the user is logged in.

Once this cookie is set, the native code can inject some javascript code
to regiter itself using its FCM (Firebase Cloud Messaging) token.